Move topNav component out of /bidder folder cuz it's used in poster page as well
