+++ client/src/pages/component-showcase.tsx
@@ -2,7 +2,6 @@ import { useState } from "react";
 
 import BottomNav from "@/components/ui/bidder/bottom-nav";
 import TaskCard from "@/components/ui/bidder/task-card";
-import TopNavtab from "@/components/ui/bidder/top-navtab";
 import BidderOfferCard from "@/components/ui/bidder-offer-card";
 import { Button } from "@/components/ui/button";
 import {
@@ -34,6 +33,7 @@ import {
 import PersonDetail from "@/components/ui/person-detail";
 import PosterTaskCard from "@/components/ui/poster/task-card";
 import ProfileTag from "@/components/ui/profile-tags";
+import TopNavtab from "@/components/ui/top-navtab";
 
 function ComponentSection({
   title,
